+++ src/main/java/ro/unibuc/careerquest/exception/UsernameTakenException.java
@@ -1,4 +1,4 @@
-package ro.unibuc.hello.exception;
+package ro.unibuc.careerquest.exception;
 
 public class UsernameTakenException extends RuntimeException {
     private static final String usernameTakenTemplate = "Username %s already taken";
